Fix: vector::reserve takes up a lot of memory
